Fixed warnings in RobWorkStudio.
git-svn-id: https://svnsrv.sdu.dk/svn/RobWork/trunk@5124 494c09c7-a67d-49d4-9938-151fb1965be8

+++ RobWorkStudio/src/rws/ImageView.cpp
@@ -51,7 +51,7 @@ void ImageView::display(const rw::sensor::Image& image) {
                     value += (((int)(255*val))&0xFF)<<8;
                     value += (((int)(255.0*val))&0xFF);
 
-                    qimage.setPixel(i,j, value);
+                    qimage.setPixel((int)i,(int)j, value);
                 }
             }
         } else {
@@ -64,7 +64,7 @@ void ImageView::display(const rw::sensor::Image& image) {
                     value += (((int)(255.0*pixel.ch[1]))&0xFF)<<8;
                     value +=  ((int)(255.0*pixel.ch[2])&0xFF);
 
-                    qimage.setPixel(i,j, value);
+                    qimage.setPixel((int)i,(int)j, value);
                 }
             }
         }
